Add files via upload
This code will create a simple web page that displays a real-time graph of the water level in the tank. The page will connect to an Arduino running the code provided in the previous step using WebSocket, and it will update the graph as new data is received from the Arduino.

Please note that you will need to modify the code to match the specific URL of your Arduino WebSocket server. You can also add additional features to the code, such as historical data storage and displaying warnings when the water level reaches certain thresholds.